requirements.txt
# 在此列出你的 Python 套件，例如：
pyserial
